Edit Recovery: switch back from form.reset() to manual setting
Rather than using HTMLFormElement.reset() to discard the Edit
Recovery data, return to how we were doing it previous to the
below change, and save and then set the old form data. This
should be identical in terms of what data is set, but it means
that the textarea value is set via the TextSelection API and so
CodeMirror will see it and work correctly.

Bug: T353616
Change-Id: I1d746b4a0e42447751e4fcb5b4b9386ac36ee649
Follow-Up: I1ec757d5ddd9f0db66496c6aaef70747d93a5c83