Edit Recovery: Add config.json to special page and postEdit
The edit recovery config was added as a package file in
Ibac3a531ccaa7bbf8d667bc0fe3e873fd0565f70 but it's also needed on
the special page, as well as in the postEdit script.

It's given the same filename because storage.js
is loaded everywhere that Edit Recovery is used.

Bug: T354167
Change-Id: Iaa0f3fb3cb798ae7a4f9e0f6259fbe6fdc3d30cd