Refactor ComplexNamespaceInputWidget to use namespaced OOUI classes
This patch refactors the ComplexNamespaceInputWidget class to use
fully qualified class names for OOUI components. It removes the manual
specification of the OOUI namespace within the class and instead relies
on the `use` declarations at the top of the PHP file. Additionally,
it adds the Exception class to the list of used namespaced classes,
preparing for explicit exception handling.

Change-Id: I3fe255f491979b63b5361da69a87c32b4fa65b5f