Edit Recovery: Improve determination of undo state
The 'undoafter' check is not sufficient, because when undoing
revisions that are not the latest or the one before, the 'undoafter'
parameter is 0, but the 'undo' parameter is not (in any undo).
To fix this, switch to just checking the latter.

Bug: T347869
Follow-Up: Icd8cc8b5d221969595c268c8c0c3ff59aef4dc80
Change-Id: Id27a04aab70da5a6e8c5df3cb6e5fdf4fc86cd16