Edit Recovery: Only delete in postEdit if data was saved
Store a temporary flag variable in sessionStorage to indicate to
the postEdit hook handler that there is data to be deleted.

This avoids the need to determine, server-side in the
Article class, whether the post-edit request has come from an
edit form that has Edit Recovery enabled (e.g. undo and edit
conflicts do not have it enabled, but the postEdit hook is still
fired after those).

Bug: T347869
Change-Id: I839d74f65138d11f77d0e8875c7ffaaa2a56c158