Skip TextContentHandler::merge3() when texts are identical
This is mostly done for performance reasons, but can even be
user-facing in rare situations. Unfortunately it's really hard
if not impossible to provide steps to reproduce it. I hope the
proposed line of code makes sense on it's own.

The setup that made me run into this edge-case is explained in
https://www.mediawiki.org/wiki/Topic:Wugee2vl8m8kqr2h
But the idea is always the same, with or without the TwoColConflict
interface: When $wgDiff3 is misconfigured wfMerge() will return
false, as if a conflict was detected. But we can easily check if
it's even possible to have a conflict (impossible when the texts are
identical) even without having access to diff3.

The relevant caller is EditPage::mergeChangesIntoContent(). We could
add the new conditional to this spot as well. But I feel like it
makes more sense to always do it, for all possible callers. The
extra identity check is extremely inexpensive, even with long texts.

Change-Id: I5b1438b1ff71590718120367eb414b8e9754236e