EditPage: Improve auto-merge in trivial non-conflict situations
This is a direct follow up to I5b1438b. The check we added there
already made the situation much better. It triggers when the user
didn't even made a change to the text. The new check(s) in this
patch trigger when the edit is based on either the latest revision
or another revision identical to the latest revision. A 3-way-merge
doesn't make sense in these situations because there aren't 3
different versions, only 2.

For reference:
* $oldContent = the base revision the edit is based on
* $myContent = the unsaved edit we are trying to save
* $yourContent = the latest revision already in the database

As in I5b1438b it's unfortunately not possible to give easy steps
to reproduce the issue. The setup that made me run into this is
explained in
https://www.mediawiki.org/wiki/Topic:Wugee2vl8m8kqr2h
This is also nice to have for performance reasons, and might help
reducing other false-positives where conflicts are detected that
aren't really conflicts.

One way to test this on dev machines is to break your $wgDiff3
config. This will make wfMerge() always return false, as if every
edit is a conflict.

Change-Id: I1a1523d1cc6fb09472cb40e41da85109a5fce851