userOptions: Support --old being the option's default value
Prior to this patch, userOptions.php couldn't be used
to change an option's value from its default to a new value,
since the script strictly expected $from to equal up_value
in user_properties, which never happens if $from is the default
value.

This patch fixes this by doing a LEFT JOIN and examining
missing user_properties rows (ie. cases when up_value is NULL).
For each such case, $from is compared with the default value
for the given user and if it matches, the option is changed.

To ensure this querying does not take long time when it is not
actually required, the new behaviour only applies
when --old-is-default is passed.

Bug: T354417
Change-Id: I95781588e9b494ef479790368e557c9182bdb6f8