Edit Recovery: use live diff view
In order to avoid inconsistencies when discarding Edit Recovery
changes when live preview is not enabled, force it to be used
when viewing changes from the Edit Recovery notification.

As Edit Recovery is a JS-only feature, there is no problem with
requiring live preview.

This refactors the normal live diff view, to move the construction
of the `<div id="wikiDiff">` element into the reusable doPreview()
function. That way, neither the normal live diff nor the new Edit
Recovery live diff need to create their own #wikiDiff elements
(they would only need to do so if they didn't want them in the
normal place).

Bug: T354494
Change-Id: Ic302ef14175b7cefa4d18ea5bc3ebcd7bd3b3b10