In ConverterRule:::parseRules log and ignore preg_split() failure
If the converter rule input text is so large that it makes preg_split()
fail, typically due to the backtracking limit, log a warning and ignore
the rule. This avoids a PHP error.

The channel and log message are similar to the one in MagicWordArray.

Iaab4849d3d7d5a62 makes errors less likely but doesn't eliminate them
entirely.

Bug: T299757
Change-Id: I8a6c6b421cd5b89cd1b80e9b417147a93793674e