SessionProvider: add ability to declare maintenance users
Adds a new method to the SessionProviderInterface that declares if the
session provide is exempt from user creation authorization.  This is
intended to support use cases where the SessionProvider is providing
access to maintenance related users external to MediaWiki itself. For
example in a supporting application to a wiki farm. This lets the
SessionProvider declare that account creation, in it's context, is
allowable even if anonymous account registration is disabled.

Bug: T345185
Change-Id: Iab3d93a3457e73eff408c1938fbdcb01d8be9a7a