user preferences: Add some information about types to the phpdoc
User options saved to the database and retrieved get converted
to strings, but defaults and set but not yet saved vaues don't.
This behavior can be confusing.

When a value equal to the default is set for a given user, that
value won't be written to the database; defaults are represented
by the lack of a database row. This behavior can also be
confusing.

Update the phpdoc of some user preferences related methods and
variables to make these things clearer.

Change-Id: I32a18b30f007c97015e6d641379ce60ab6ac7ed7