updateCollation: Simplify and redo how batching works
Long time ago, when we changed collation way more often and the HDD was
the norm for databases, indexing on cl_collation helped us speed up the
updates and minimize user impact. That's not the case anymore and on top
of that, we now have feature of copying the table via setting
--target-table and there is no need to run the schema change super fast.

Also, categorylinks table is quite large (in Commons it has reached
210GB, comparable to enwiki's revision table) and an extra index like
that is quite taxing on the infra.

So let's just do whatever other scripts do, go through all rows in batches
and take advantage of cl_from index instead. This is similar to what
migrateLinksTable does.

Bug: T342854
Change-Id: Ie4dd91ee29308c980ec0b9b7ee684cb175ffca43