Reword documentation for user.user_is_temp to clarify purpose
Clarify that this column is for external tools that only have
access to the database, and that it should not be used within
MediaWiki.

This is to avoid there being two different ways to look up if a
user is temporary. Parsing the name is preferred within MediaWiki,
because it can be done directly from the client and it corresponds
to how other user types are determined (e.g. external users).

The need for this column indicates technical debt, specifically
the need for a user type system, discussed in T336176. It is
considered temporary and replaceable by whatever system is
designed.

Bug: T333223
Change-Id: I94ec08daf78e76273ca055d21f5cea85021490c4