title: Merge largely duplicated Title::getEditNotices code
This was the almost exact same piece of code repeated 3 times, with
the only difference being the message key and one of the CSS class
names.

This duplication is very old. Since commit d894b949 from 2011 the
basic logic is repeated 3 times. Since I4ad5e16 from 2015 the Html
class is used, which made the duplication much more visible.

Change-Id: Ic8201d3a3f38316e566830293c0b4ddbf63f09be