stats: allow passing a single/multiple namespaces to copyToStatsdAt()
There are a few places in MediaWiki code where we copy the same
measurement to multiple statsd namespaces.  This patch enables StatsLib
to take an array of namespaces and pass the single measurement call to
each configured statsd namespace.

Refactors handling of $statsdNamespaces to BaseMetric to consolidate
validation logic.

Adds tests.

Bug: T355361
Change-Id: Ib407d0d33dbcb5ae9b77518e01ee0917217685b4