Edit Recovery: Delete the data-saved sessionStorage flag
When ER is used on an edit page, it sets a sessionStorage variable
to let the postEdit hook know that it should delete the saved ER
data. But ER isn't always loaded on the edit page, and when e.g.
saving an undo we don't want to delete the stored data. The
data-saved flag lasts for five minutes, so if an undo is performed
in that time then the ER data was also being deleted.

This moves the ER load check to be within the RL module (so the
module is now loaded on every edit form), and when
ER is *not* loaded, it deletes the sessionStorage variable.

Bug: T347869
Change-Id: I529a9636567d318ae9bc61f46469c5d45881f46b