upload: Remove a duplicate strpos call in UploadBase
Just use the result from one call in both places.

I'm also re-arranging the code for readability. This is quite
critical here. A file name like ".htaccess" where the very first
character is a dot but no other dot follows must be considered a
filename without an extension. I hope this is more visible with
the `> 0`.

Change-Id: I24179de62c3f4443effe8a4ebd089a3f77fd84e3