maintenance: Fix broken strpos in findDeprecated script
The idea here was obviously to speed up the process by skipping files
that don't contain the substring "@deprecated" anywhere. Only these
files are parsed and traversed – which can be expensive.

The problem is that PHP's strpos() function never returns -1. It
returns false.

This patch doesn't change what the script does. It just runs faster.

Change-Id: I95a5a0fd3e024ec4132f53d770e3f61031d81250