Fix confusing strpos in LanguageConverter::markNoConversion
This code was added in 2005 via commit ed8d7084, see
https://phabricator.wikimedia.org/rMWed8d7084
Luckily it's not terribly broken (i.e. the bug shouldn't cause any
trouble in real-world scenarios) but it's not correct either.

1. How it currently behaves: Note that strpos() in PHP can return 0
when a match was found at the start of the string, and false when
nothing was found. This was not differentiated in the old code.

 if ( strpos( $text, '-{' ) || strpos( $text, '}-' ) )

The second strpos is mostly fine because the closing `}-` can't
appear at the start of the string anyway. But the first strpos works
only if the opening `-{` is *not* at the start of the string, but
later in the string. Luckily this doesn't cause much trouble because
it's very likely that a snippet of wikitext that starts with `-{`
also contains `}-`.

2. How it was (probably) meant to behave: The method surrounds a
snippet of wikitext with `-{…}-` to *disable* conversions. This
doesn't make sense when the snippet already contains `-{…}-`
sections that *should* be converted.

The only possible change in behavior is when a string starts with
`-{` but doesn't contain `}-`. In reality such strings don't exist,
as far as I can see. And even if, it feels more correct to not
surround a string with such broken syntax with more (broken) syntax.

Change-Id: Ifdec32a52db3c5a94768692e957da962b6dd00b3