rdbms: Remove $domain support from ::getDBFromRecency
This method is a shallow component [1] which is an anti-pattern in
software design (yes, I'm aware I introduced that method). We should
avoid expanding its interface further and focus on the most common
usecase. Since it's not doing much under the hood, it's okay to just
copy-paste the logic into a method in UserGroupManager.

This partially reverts I9e2de1cc84d12443.

[1] Original idea from "Philosophy of software design" book but you can
find short TLDRs in the internet, like:
https://vladimirzdrazil.com/posts/deep-shallow-modules/

Bug: T355813
Change-Id: I8cac972ecc9e962f6605a8feaa24a9e00af07b66