ParserTestRunner: Bug fix that skipped selser tests in CI
This only impacts PHPUnit testing (not the CLI test runner)
since we init numchanges to 0 because PHPUnit cannot run dynamic
tests. We instead install a single selser-auto-composite test
which then runs *all* selser auto-edit tests in a single run.

This bug affected that run since the reset to 20 failed with the
use of "??" instead of "?:"

Change-Id: Id81e2f4331ce73145389e717b7c78f49b29743bf