i18n: Change default label of Special:Log, and entry in drop-down to 'Main'
This is, and has always been, a filtered list even of public logs, so let's
change the label for now to that it's less misleading. There are many, many
difficulties with the log pages' design but there's no apparent appetite to
tackle this in the short- or even medium-term, so a minor fix might help to
be slightly less bewilderingly wrong.

Bug: T237729
Change-Id: I103e5ffd4751d6cae4c4f3b654c51eb20cfda471