ParserTestRunner: Fix behavior of bare `parsoid-compatible` file option
The intent was that a `parsoid-compatible` flag without specified modes
would enable all of the parsoid modes.  However, the value of
`$fileOptions['parsoid-compatible']` in that case is '', and the
test was looking specifically for a null or false value, not a falsey
one.

This exposed a bug where, because the LinkCache was destroyed and reset,
the calls to Title::clearCaches() didn't have the expected effect, since
the LinkCache held by Title was different from the one held by the
PageStore.  This caused the wrong page IDs to be associated with titles,
which then in turn caused the Disambiguator extension (which looks up
page properties by page id) to return the wrong results.

Depends-On: Ib65dd0adb472da53823c07af5991a140374501e4
Depends-On: Id9d6b3ab9d168048f46b1ab8783d06a7cbc9e6d4
Change-Id: Ifca13393c3bbec27c23cbdc311d4550fbccf21ca