composer: Allow unreleased composer versions
If you try to do `composer update` using a git checkout of a Composer
branch, it fails because "@package_version@" is less than 2.0.0.

So in that case, use Composer::getVersion() which falls back to
Composer::SOURCE_VERSION. The method doesn't exist on very old versions
of Composer, so we can't use it unconditionally.

Change-Id: I53ea01c2f992499a25ba851bd8ed3593c9a13711