userOptions: Add support for --delete-defaults
As part of using conditional defaults in Echo, database
rows matching the current defaults should be deleted. This
is easiest to do by a maintenance script that iterates through
all rows for that option

While this is slow, there is no other way to process conditional
defaults. The slowness should not be an issue, because the script
will be ran as a one-off cleanup action.

Bug: T355367
Change-Id: Id60856f9942a06dc494e539f488ce3789353f88a