specials: Remove null comments from formatter on Special:ProtectedPages
Due to missing information for old protections, the comment from the log
table is null in the result set and since 6f1f783c a warning is emitted.

Fake the result set without the null rows, needs to change the key value
for the $formattedComments array as the offset no longer fits.

Bug: T356337
Change-Id: If2e656cbd4870ec7617c6fbb4e34a103413c8a35