Load Filepage.css when previewing File pages
I’m not sure whether the condition should be $this->preview or
$this->formtype === 'preview', but the latter feels slightly more
correct to me after looking through uses of both. Similarly, I’m not
sure if the namespace should be checked on $this->getContextTitle() or
$this->mTitle, but I think the latter fits better, both because the
preview seems to get parsed with $this->mTitle (see doPreviewParse())
and because $this->mTitle->getNamespace() is used for more namespace
comparisons elsewhere (e.g. when checking whether a user is editing
their own user or talk page).

Bug: T356505
Change-Id: I182b264b6b23887c05817178b76011a5cb978575