WebRequest: Remove 'deprecated' mark from unstable Telemetry pointer
While not formally specified in the Stable interface policy, it is
imho not recommended to deprecate a method in a stable class with a
pointer to a method in an `@unstable` class. This sends a mixed signal
to developers.

Keeping this as wrapper method is harmless either way, but I suggest
removing this so that we do not create send automatic nudges to
contributors to change existing code (yet).

Bug: T346211
Change-Id: Iddf8f286e3ffdfce32a8fc2cc95a4e1d4893e146