Clarify that partial action blocks block an action sitewide
Partial blocks may block editing particular pages and/or performing
certain actions. If both are specified in one block, the actions
are blocked sitewide, not just on the specified pages. Make this
clear in the help text on Special:Block.

Also update the help page link in ipb-sitewide-help to the relevant
section, to help clarify which actions are blocked.

Bug: T242541
Change-Id: If7d0f11dc121495e3e3e9ec4afe6a6fa15456f2c