[SpecialVersion] Fix double-escaping in {{CURRENTVERSION}}
Phan correctly catches a case of double-escaping here: the output of
Special::getVersion() is used as the return value for the
CoreMagicVariables implementation of {{CURRENTVERSION}} which means it
is wikitext (not escaped HTML).

No practical consequence since the only thing being escaped is a SHA
hash and the localized parentheses, neither of which are likely to contain
characters which are going to trigger the extra escape.

Change-Id: I9f125d56b9d143f2a0baea3da8bbb92b38317537