Allow filter: in inline CSS.
This was banned because it could be used to load other files,
including potentially local files, in IE9 and earlier. This
browser is no longer relevant. Wikimedia sites stopped supporting
the needed TLS versions for that browser 4 years ago.

Modern browsers have redefined filter to mean something different.
Generally the new filter is perfectly safe as long as we ban the url()
function which we do.

For context on why it was originally banned, see
https://static-codereview.wikimedia.org/MediaWiki/66990.html

Bug: T308160
Change-Id: Ic94f499dfe66e3cce12496893d0ecbee006bd243