Use ESLint "sourceType": "commonjs" for package modules
When using MediaWiki package modules, `var`/`let` etc. statements at
top-level define local variables rather than global variables, unlike
normal MediaWiki modules (and unlike <script> code in the browser),
and like CommonJS modules (although we don't follow the whole spec).

We used two different ways to silence ESLint warnings about that:
* { "rules": { "no-implicit-globals": "off" } }
* { "env": { "commonjs": true } }

Recently, a new sheriff arrived in town:
* { "parserOptions": { "sourceType": "commonjs" } }

I can't figure out which ESLint version introduced it, but
it seems to work as of 8.54.0, which we're using currently.
It's a part of a new config system they're working on, but
it seems we can use it already despite using the old one.

Related:
https://eslint.org/blog/2022/08/new-config-system-part-2/#reimagined-language-options
https://eslint.org/docs/latest/use/configure/configuration-files-new#configuring-the-javascript-source-type

Change-Id: Iac2824f012cbc2c05c3c26d69d860d9b6dc45be6