block: Include bt_auto in acquireTarget() conditions
When searching for an existing block_target row to use for a new block
of an IP address, include bt_auto in the conditions. Otherwise, if there
is an existing autoblock, the new block will fail due to a mismatch of
bt_count, leaking information about the private IP address.

Bug: T357366
Change-Id: If873ca590aadd29ab0b9d672a99438d70c4292d9