Edit Recovery: Switch recovery message language based on $wgEditSubmitButtonLabelPublish
When an editing session is recovered but there's been subsequent editing
on the page, a message is shown to highlight this fact to the user.
This change is to modify this message based on the value of
$wgEditSubmitButtonLabelPublish (save vs publish).

This also fixes up the spacing between the sentences, so that the spaces
and emphasis are part of the messages rather than hardcoded in the
module.

Bug: T351435
Change-Id: Ie66b9b5fa9a58b37d9fcc818cc176dc2c0c710b5