CodexModule: use CodexModule to load the full Codex library
This patch provides a way to load the entire Codex library in
CodexModule. This is needed when a user wants to load the
entire Codex library versus a subset module that uses
treeshaking.
- Added a new method and option to load the full Codex library
- Converted the module definitions of @wikimedia/codex and
codex-styles. The `@wikimedia/codex` module loads the entire
Codex library in CodexModule and does not utilize the  manifest
files. We removed the themeStyles map from the `codex-style`
module definition; and added a themeStyle map in the CodexModule
to ensure that styles are loaded based on the specified theme
and direction.

Bug: T355840
Change-Id: I3b4158171c69427cdcb7c705712bfd55cde81d42