Codex: Add experimental theme
This is still experimental and not necessarily stable, but adding this
for now so that skins can experiment with it.

A skin can select this theme by setting the SkinCodexThemes entry for
its skin to "experimental".

Also rework CodexModule::getManifestFilePath() to be array-based, now
that we have more than two themes to handle there.

Bug: T356542
Change-Id: I6791d4d8c6686845181af7bc79f6eecc1e601add