logging: Fix LogFormatterTestCase from throwing with CheckUser
`CheckUserPrivateEventLogFormatterTest::testLogDatabaseRows()` calls
`LogFormatterTestCase::doTestLogFormatter()` which overrides UserFactory
with a mock which will eventually call `BlockManager::getBlock()` which
tries to create a legacy user object with `newFromUserIdentity()`, but
the overriden UserFactory mock doesn't mock the method `newFromUserIdentity()`,
so will return a NULL user object.

With the NULL user object returned above and CentralAuth enabled, the
`onGetUserBlock()` hook in CA will fire and try to act on a NULL user object
causing the `CentralAuthUser::getInstanceByName()` method in I2c7b4ceb56463e88
to throw as if we're dealing with a user object having an invalid username.

In order to fix this, this patch also mocks the `newFromUserIdentity()` method
in the mocked UserFactory service that gets overridden in test. As a result,
with this patch, the failing tests in CheckUser extension should begin
passing again.

Bug: T343963
Change-Id: I11937d0d53b3914d7f6bd1483404a04b74ac26c9