watchlist: When re-watching a page, add semicolon
If usenewrc is off and watchlistunwatchlinks is on, and a page with
watchlist expiry is unwatched via the ChangesList watch link, then
rewatched by the same link, the customary semicolon separator is not
added between the article and history link.

So, fix that. Try to match the PHP logic in ChangesList::getTimestamp().

If usenewrc is on, the mw-changeslist-separator separator will not be
found, so the code will have no effect, as intended.

Also, remove the spaces which wrap the expiry icon.

Bug: T266747
Change-Id: If53818fbaa105448aad46c6b6571d3041352306d