Don't double-wrap headings when using DiscussionTools
Discussion Tools runs *before* this stage runs, and so we end up
wrapping headings which have already been wrapped by discussion tools.
Check for an existing wrapper to avoid this.

In the future, we will probably add a new post-cache transform hook
which is at the very *end* of the pipeline, instead of in the middle,
to avoid this sort of ordering dependency between extensions and core.

Bug: T357826
Change-Id: I8cd28a3b42e55844be1258d639e605862952806f